Added Clover Patch block. It still needs to spawn in the world and be able to spread, probably like grass. Bonemeal should also make it grow.
Added Palm Sapling and tree generation. The tree looks bad so a custom "tree shape" is going to be needed.

Added custom leaf block so that it burns.

Added Four-Leaf Clover item. It will be a potion ingredient.

Fixed leaf color issue. It probably needs to go in a client event class, but it's in the main for now.