Added Patchouli. Currently just has test items in it.
Clover Patch no longer has an item registered for it. It still doesn't spread, but now it drops Clover with a chance for Four-Leaf Clovers.

Added Clover Lawn and Clover Lawn (Full). The latter has a terrible name.

Added event for Dolphins to convert an item into something else when they play with it. It turns Dragon Scales into Diamond Blocks, which is obviously a placeholder.

Added Clover item. It can be used to make Clover Lawn. It still needs to be able to be placed in a Composter.

Attempted to add/fix some tag issues. I think I got them all, except for crafting Palm Planks. And logs can't be "shift-clicked" into a furnace.